Updtaed stuctures
